A watchlist scheme to allow shared clauses and make better use of cache

+++ mtl/Alg.h
@@ -40,7 +40,7 @@ static inline void remove(V& ts, const T& t)
     uint32_t j = 0;
     for (; j < ts.size() && ts[j] != t; j++);
     assert(j < ts.size());
-    for (; j < ts.size()-1; j++) ts[j] = ts[j+1];
+    for (; j < (uint16_t)(ts.size()-1); j++) ts[j] = ts[j+1];
     ts.pop();
 }
